backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
celery==5.4.0
redis==5.1.1
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic-settings==2.5.2
python-multipart==0.0.12
groq==0.11.0
google-generativeai==0.8.3
yt-dlp==2024.11.4
opencv-python-headless==4.10.0.84
librosa==0.10.2
soundfile==0.12.1
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
